move SerialScale to KokkosBlas
