+++ src/batched/dense/impl/KokkosBatched_Gemm_Serial_Internal.hpp
@@ -6,7 +6,7 @@
 #include "KokkosBatched_Util.hpp"
 
 #include "KokkosBatched_Set_Internal.hpp"
-#include "KokkosBatched_Scale_Internal.hpp"
+#include "KokkosBlas1_serial_scal_impl.hpp"
 
 #include "KokkosBatched_InnerGemmFixC_Serial_Impl.hpp"
 
@@ -43,7 +43,7 @@ KOKKOS_INLINE_FUNCTION int SerialGemmInternal<Algo::Gemm::Unblocked>::invoke(
   if (beta == zero)
     SerialSetInternal ::invoke(m, n, zero, C, cs0, cs1);
   else if (beta != one)
-    SerialScaleInternal::invoke(m, n, beta, C, cs0, cs1);
+    KokkosBlas::Impl::SerialScaleInternal::invoke(m, n, beta, C, cs0, cs1);
 
   if (alpha != zero) {
     if (m <= 0 || n <= 0 || k <= 0) return 0;
@@ -83,7 +83,7 @@ KOKKOS_INLINE_FUNCTION int SerialGemmInternal<Algo::Gemm::Blocked>::invoke(
   if (beta == zero)
     SerialSetInternal ::invoke(m, n, zero, C, cs0, cs1);
   else if (beta != one)
-    SerialScaleInternal::invoke(m, n, beta, C, cs0, cs1);
+    KokkosBlas::Impl::SerialScaleInternal::invoke(m, n, beta, C, cs0, cs1);
 
   if (alpha != zero) {
     if (m <= 0 || n <= 0 || k <= 0) return 0;
