+++ src/batched/dense/impl/KokkosBatched_Gemv_Serial_Internal.hpp
@@ -6,7 +6,7 @@
 #include "KokkosBatched_Util.hpp"
 
 #include "KokkosBatched_Set_Internal.hpp"
-#include "KokkosBatched_Scale_Internal.hpp"
+#include "KokkosBlas1_serial_scal_impl.hpp"
 
 #include "KokkosBatched_InnerMultipleDotProduct_Serial_Impl.hpp"
 
@@ -41,7 +41,7 @@ KOKKOS_INLINE_FUNCTION int SerialGemvInternal<Algo::Gemv::Unblocked>::invoke(
   if (beta == zero)
     SerialSetInternal ::invoke(m, zero, y, ys0);
   else if (beta != one)
-    SerialScaleInternal::invoke(m, beta, y, ys0);
+    KokkosBlas::Impl::SerialScaleInternal::invoke(m, beta, y, ys0);
 
   if (alpha != zero) {
     if (m <= 0 || n <= 0) return 0;
@@ -80,7 +80,7 @@ KOKKOS_INLINE_FUNCTION int SerialGemvInternal<Algo::Gemv::Blocked>::invoke(
   if (beta == zero)
     SerialSetInternal ::invoke(m, zero, y, ys0);
   else if (beta != one)
-    SerialScaleInternal::invoke(m, beta, y, ys0);
+    KokkosBlas::Impl::SerialScaleInternal::invoke(m, beta, y, ys0);
 
   if (alpha != zero) {
     if (m <= 0 || n <= 0) return 0;
