+++ src/batched/dense/impl/KokkosBatched_Trsv_Serial_Internal.hpp
@@ -6,7 +6,7 @@
 #include "KokkosBatched_Util.hpp"
 
 #include "KokkosBatched_Set_Internal.hpp"
-#include "KokkosBatched_Scale_Internal.hpp"
+#include "KokkosBlas1_serial_scal_impl.hpp"
 
 #include "KokkosBatched_InnerTrsm_Serial_Impl.hpp"
 #include "KokkosBatched_Gemv_Serial_Internal.hpp"
@@ -44,7 +44,8 @@ SerialTrsvInternalLower<Algo::Trsv::Unblocked>::invoke(
   if (alpha == zero)
     SerialSetInternal::invoke(m, zero, b, bs0);
   else {
-    if (alpha != one) SerialScaleInternal::invoke(m, alpha, b, bs0);
+    if (alpha != one)
+      KokkosBlas::Impl::SerialScaleInternal::invoke(m, alpha, b, bs0);
     if (m <= 0) return 0;
 
     for (int p = 0; p < m; ++p) {
@@ -81,7 +82,8 @@ KOKKOS_INLINE_FUNCTION int SerialTrsvInternalLower<Algo::Trsv::Blocked>::invoke(
   if (alpha == zero)
     SerialSetInternal::invoke(m, zero, b, bs0);
   else {
-    if (alpha != one) SerialScaleInternal::invoke(m, alpha, b, bs0);
+    if (alpha != one)
+      KokkosBlas::Impl::SerialScaleInternal::invoke(m, alpha, b, bs0);
     if (m <= 0) return 0;
 
     /// case GPU: team size is large and blocksize (mb,nb) is small
@@ -137,7 +139,8 @@ SerialTrsvInternalUpper<Algo::Trsv::Unblocked>::invoke(
   if (alpha == zero)
     SerialSetInternal::invoke(m, zero, b, bs0);
   else {
-    if (alpha != one) SerialScaleInternal::invoke(m, alpha, b, bs0);
+    if (alpha != one)
+      KokkosBlas::Impl::SerialScaleInternal::invoke(m, alpha, b, bs0);
     if (m <= 0) return 0;
 
     ValueType *KOKKOS_RESTRICT b0 = b;
@@ -172,7 +175,8 @@ KOKKOS_INLINE_FUNCTION int SerialTrsvInternalUpper<Algo::Trsv::Blocked>::invoke(
   if (alpha == zero)
     SerialSetInternal::invoke(m, zero, b, bs0);
   else {
-    if (alpha != one) SerialScaleInternal::invoke(m, alpha, b, bs0);
+    if (alpha != one)
+      KokkosBlas::Impl::SerialScaleInternal::invoke(m, alpha, b, bs0);
     if (m <= 0) return 0;
 
     InnerTrsmLeftUpperUnitDiag<mbAlgo> trsm_u(as0, as1, bs0, 0);
