+++ src/batched/dense/impl/KokkosBatched_Trsv_Team_Internal.hpp
@@ -6,7 +6,6 @@
 #include "KokkosBatched_Util.hpp"
 
 #include "KokkosBatched_Set_Internal.hpp"
-#include "KokkosBatched_Scale_Internal.hpp"
 
 #include "KokkosBatched_InnerTrsm_Serial_Impl.hpp"
 #include "KokkosBatched_Gemv_Team_Internal.hpp"
