+++ src/blas/impl/KokkosBlas3_trsm_impl.hpp
@@ -75,7 +75,7 @@ int SerialTrsmInternalLeftLowerConj(const bool use_unit_diag, const int m,
     KokkosBatched::SerialSetInternal ::invoke(m, n, zero, B, bs0, bs1);
   else {
     if (alpha != one)
-      KokkosBatched::SerialScaleInternal::invoke(m, n, alpha, B, bs0, bs1);
+      KokkosBlas::Impl::SerialScaleInternal::invoke(m, n, alpha, B, bs0, bs1);
     if (m <= 0 || n <= 0) return 0;
 
     for (int p = 0; p < m; ++p) {
@@ -114,7 +114,7 @@ int SerialTrsmInternalLeftUpperConj(const bool use_unit_diag, const int m,
     KokkosBatched::SerialSetInternal ::invoke(m, n, zero, B, bs0, bs1);
   else {
     if (alpha != one)
-      KokkosBatched::SerialScaleInternal::invoke(m, n, alpha, B, bs0, bs1);
+      KokkosBlas::Impl::SerialScaleInternal::invoke(m, n, alpha, B, bs0, bs1);
     if (m <= 0 || n <= 0) return 0;
 
     ValueType* KOKKOS_RESTRICT B0 = B;
