+++ src/sparse/impl/KokkosSparse_spmv_bsrmatrix_impl.hpp
@@ -542,7 +542,6 @@ struct BsrMatrixSpMVTensorCoreDispatcher {
 #include "KokkosBatched_Gemv_TeamVector_Internal.hpp"
 #include "KokkosBatched_Gemm_Serial_Internal.hpp"
 #include "KokkosBatched_Gemm_TeamVector_Internal.hpp"
-#include "KokkosBatched_Scale_Internal.hpp"
 #include "KokkosKernels_ExecSpaceUtils.hpp"
 
 #include "KokkosBlas2_team_gemv_spec.hpp"
